Refactor stateful data widgets
