+++ lib/stateful_data/stateful_data.dart
@@ -1,4 +1,4 @@
 library stateful_data;
 
 export 'stateful_data_model.dart';
-export 'loadable_uic_view.dart';
+export 'stateful_data_view.dart';
