+++ test/stateful_data_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:uic/stateful_data/loadable_uic.dart';
+import 'package:uic/stateful_data/stateful_data.dart';
 
 void main() {
   setUp(() {
